refactor: 글쓰기 페이지 코로케이션 적용 및 목데이터 제거
